Double quote MySQL strings in --alter flag value
Values inside the MySQL ALTER command will often be single-quoted. Change these to be doubly-quoted strings so they are interpreted correctly within the singly-quoted --alter value.
The --alter value cannot be double quoted because it contains expressions in backticks that are not meant to be evaluated by the shell.

Add a test to ensure that the single quotes are not being nested within the command.

Fixes PagerDuty/pt-osc#12