Warn when executing ALTER commands without PtOscMigration
If an ALTER command is being called on the MysqlPtOscAdapter without coming from a PtOscMigration, this usually means that commands will be queued up in the commands hash and never actually executed. This results in silent failure.

To make this failure more apparent, issue a warning when add_command is called and PtOscMigration does not appear in the stack trace.